Meson

+++ freeciv/prepare_freeciv.sh
@@ -1,6 +1,8 @@
 #!/bin/bash
 
-MESON_VER=0.57.2
+# Either version number to install, or "yes" to use system meson.
+# Empty to use autotools
+MESON_VER="yes"
 
 DIR="$(cd "$(dirname "${BASH_SOURCE[0]}")" >/dev/null && pwd)"
 cd "${DIR}"
@@ -28,16 +30,18 @@ if ! ./apply_patches.sh ; then
 fi
 
 if test "$MESON_VER" != "" ; then
-  rm -Rf meson-install && mkdir -p meson-install
+  if test "$MESON_VER" != "yes" ; then
+    rm -Rf meson-install && mkdir -p meson-install
 
-  ( cd meson-install
+    ( cd meson-install
 
-    wget "https://github.com/mesonbuild/meson/releases/download/${MESON_VER}/meson-${MESON_VER}.tar.gz"
-    tar xzf meson-${MESON_VER}.tar.gz
-    ln -s "meson-${MESON_VER}/meson.py" meson
-  )
+      wget "https://github.com/mesonbuild/meson/releases/download/${MESON_VER}/meson-${MESON_VER}.tar.gz"
+      tar xzf meson-${MESON_VER}.tar.gz
+      ln -s "meson-${MESON_VER}/meson.py" meson
+    )
 
-  export PATH="$(pwd)/meson-install:$PATH"
+    export PATH="$(pwd)/meson-install:$PATH"
+  fi
 
   mkdir -p build
